EO-885 - Rename id variable and helper function names
